Added uint64_t size support to character arrays.
Character arrays should now be able to be sent with sizes larger
then the previous limit of 4GB.

+++ include/mrnet/DataElement.h
@@ -93,7 +93,7 @@ class DataElement {
 
  private:
     DataType type;
-    uint32_t array_len;
+    uint64_t array_len;
     bool destroy_data;
 };
